upfated gifs for landing page n data
