+++ client/src/components/LandingPageComponents/LandingCarousel/LandingCarousel.jsx
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {Carousel } from 'react-bootstrap';
-import features from '../../../data/production.features';
+import features from '../../../data/productionFeatures';
 
 
 function LandingCarousel() {
@@ -23,7 +23,7 @@ function LandingCarousel() {
                             className="d-block w-100"
                             src={feature.src}
                             alt={feature.title}
-                            style={{width: '70%', maxWidth: 650}}
+                            style={{width: '67%', maxWidth: 650}}
                         />
                         
                         <p className='text-center mt-3'> {feature.details} </p>
