+++ client/src/data/productionFeatures.js
@@ -1,0 +1,41 @@
+const features = [
+   
+    {
+        src: '/assets/images/landingimg/watercooler.png',
+        title: 'Everyone Loves WaterCooler Moments',
+        details:'Times when we strike up casual conversations and learn interesting things about one another.'
+    },
+    {
+        src: '/assets/images/landingimg/gui.gif',
+        title: 'It Is Playful! So Much Fun! 🥳',
+        details:'Walkaround the room, say hi to people and invite them to social space.'
+    },
+    {
+        src: '/assets/images/landingimg/chat.gif',
+        title: 'Creat A Rooom, Pick RoomStyle & Invite Friends',
+        details:'Think of this as you host a house party. You set the theme and decide who to invite.'
+    },
+    {
+        src: '/assets/images/landingimg/videochats.png',
+        title: 'Video Chats in The Social Space',
+        details:'Want to have a more private conversation with someone(s)? Invite them to social space for video chats'
+    },
+    {
+        src: '/assets/images/landingimg/gathering.gif',
+        title: 'Interact & Socialize Just Like Around WaterCooler',
+        details:'Come and go as you please. If this room does not interest you, check out another one.'
+    },
+    {
+        src: '/assets/images/landingimg/friends.png',
+        title: 'Connect With More Friends',
+        details:'It is easier to bond with people with whom you share common interests.'
+    },
+    {
+        src: '/assets/images/landingimg/chatspop.gif',
+        title: 'Yes! It Is Mobile Friendly 😇',
+        details:''
+    },
+  
+];
+
+export default features;
